Move lockfile related functions to Lockfile module
